+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"encoding/json"
-	"encoding/xml"
 	"flag"
 	"fmt"
 	"log"
@@ -12,8 +11,8 @@ import (
 	"time"
 
 	"github.com/hornbill/color"
-	"github.com/hornbill/goApiLib"
-	"github.com/hornbill/goHornbillHelpers"
+	apiLib "github.com/hornbill/goApiLib"
+	hornbillHelpers "github.com/hornbill/goHornbillHelpers"
 )
 
 var (
@@ -22,6 +21,12 @@ var (
 
 func main() {
 	parseFlags()
+
+	//-- If configVersion just output version number and die
+	if configVersion {
+		fmt.Printf("%v \n", version)
+		return
+	}
 	logTimeNow = time.Now().Format("20060102150405.999999")
 
 	//Does endpoint exist?
@@ -43,7 +48,7 @@ func main() {
 
 	//Ask if we want to delete before continuing
 	fmt.Println("")
-	fmt.Println("===== Hornbill Cleaner Utility v" + toolVer + " =====")
+	fmt.Println("===== Hornbill Cleaner Utility v" + version + " =====")
 	if !cleanerConf.CleanRequests && !cleanerConf.CleanAssets && !cleanerConf.CleanUsers {
 		color.Red("No entity data has been specified for cleansing in " + configFileName)
 		return
@@ -75,11 +80,11 @@ func main() {
 
 	if !configSkipPrompts {
 		fmt.Println("Are you sure you want to permanently delete these records? (yes/no):")
-		if hornbillHelpers.ConfirmResponse("") != true {
+		if !hornbillHelpers.ConfirmResponse("") {
 			return
 		}
 		color.Red("Are you absolutely sure? Type in the word 'delete' to confirm...")
-		if hornbillHelpers.ConfirmResponse("delete") != true {
+		if !hornbillHelpers.ConfirmResponse("delete") {
 			return
 		}
 		espLogger("Confirmation Prompts Accepted", "info")
@@ -89,9 +94,6 @@ func main() {
 
 	//Log the config
 	logConfig()
-
-	maxResults = getMaxRecordsSetting()
-
 	processRequests()
 	processAssets()
 
@@ -105,10 +107,6 @@ func main() {
 			}
 
 		}
-	}
-	//Reset maxResultsAllowed system setting back to what it was before the process ran
-	if resetResults && !configDryRun {
-		setMaxRecordsSetting(maxResults)
 	}
 }
 
@@ -209,16 +207,8 @@ func processRequests() {
 			if !configDryRun {
 				espLogger("Number of Requests to delete: "+strconv.Itoa(requestCount), "debug")
 				color.Green("Number of Requests to delete: " + strconv.Itoa(requestCount))
-				if maxResults > 0 && requestCount > maxResults {
-					resetResults = true
-					//Update maxResultsAllowed system setting to match number of records to be deleted
-					setMaxRecordsSetting(requestCount)
-				}
 				processEntityClean("Requests", configBlockSize)
-			} else {
-				//output to log
 			}
-
 		} else {
 			requestCount = getRecordCount("h_itsm_requests")
 			if requestCount > 0 {
@@ -228,13 +218,7 @@ func processRequests() {
 				totalBlocks = int(math.Ceil(requestBlocks))
 				espLogger("Request Blocks: "+strconv.Itoa(int(requestBlocks)), "debug")
 				espLogger("Total Blocks: "+strconv.Itoa(totalBlocks), "debug")
-
 				espLogger("Number of Requests to delete: "+strconv.Itoa(requestCount), "debug")
-				if maxResults > 0 && requestCount > maxResults {
-					resetResults = true
-					//Update maxResultsAllowed system setting to match number of records to be deleted
-					setMaxRecordsSetting(requestCount)
-				}
 				processEntityClean("Requests", configBlockSize)
 			} else {
 				espLogger("There are no requests to delete.", "debug")
@@ -253,17 +237,11 @@ func processAssets() {
 			assetBlocks := float64(assetCount) / float64(configBlockSize)
 			totalBlocks = int(math.Ceil(assetBlocks))
 			espLogger("Number of Assets to delete: "+strconv.Itoa(assetCount), "debug")
-			if maxResults > 0 && assetCount > maxResults {
-				resetResults = true
-				//Update maxResultsAllowed system setting to match number of records to be deleted
-				setMaxRecordsSetting(assetCount)
-			}
 			processEntityClean("Asset", configBlockSize)
 		} else {
 			espLogger("There are no assets to delete.", "debug")
 			color.Red("There are no assets to delete.")
 		}
-
 		if !configDryRun {
 			assetLinkCount := getRecordCount("h_cmdb_links")
 			if assetLinkCount > 0 {
@@ -271,11 +249,6 @@ func processAssets() {
 				assetLinkBlocks := float64(assetLinkCount) / float64(configBlockSize)
 				totalBlocks = int(math.Ceil(assetLinkBlocks))
 				espLogger("Number of Asset Links to delete: "+strconv.Itoa(assetLinkCount), "debug")
-				if maxResults > 0 && assetLinkCount > maxResults {
-					resetResults = true
-					//Update maxResultsAllowed system setting to match number of records to be deleted
-					setMaxRecordsSetting(assetLinkCount)
-				}
 				processEntityClean("AssetsLinks", configBlockSize)
 			} else {
 				espLogger("There are no asset links to delete.", "debug")
@@ -292,35 +265,8 @@ func parseFlags() {
 	flag.StringVar(&configInstance, "instance", "", "ID of the instance (case sensitive)")
 	flag.BoolVar(&configDryRun, "dryrun", false, "DryRun mode - outputs record keys to log for review without deleting anything")
 	flag.BoolVar(&configSkipPrompts, "justrun", false, "Set to true to run the cleanup without the prompts")
+	flag.BoolVar(&configVersion, "version", false, "Returns the tool version")
 	flag.Parse()
-}
-
-//setMaxRecordsSetting - takes integer, updates maxResultsAllowed system setting to value
-func setMaxRecordsSetting(newMaxResults int) bool {
-	espXmlmc.OpenElement("option")
-	espXmlmc.SetParam("key", "api.xmlmc.queryExec.maxResultsAllowed")
-	espXmlmc.SetParam("value", strconv.Itoa(newMaxResults))
-	espXmlmc.CloseElement("option")
-	browse, err := espXmlmc.Invoke("admin", "sysOptionSet")
-	if err != nil {
-		color.Red("Call to sysOptionSet for api.xmlmc.queryExec.maxResultsAllowed data failed.")
-		espLogger("Call to sysOptionSet for api.xmlmc.queryExec.maxResultsAllowed data failed.", "error")
-		return false
-	}
-	var xmlRespon xmlmcResponse
-	err = xml.Unmarshal([]byte(browse), &xmlRespon)
-	if err != nil {
-		color.Red("Unmarshalling of data for sysOptionGet for api.xmlmc.queryExec.maxResultsAllowed failed.")
-		espLogger("Unmarshalling of data for sysOptionGet for api.xmlmc.queryExec.maxResultsAllowed failed.", "error")
-		return false
-	}
-	if xmlRespon.MethodResult != "ok" {
-		color.Red("sysOptionSet was unsuccessful: " + xmlRespon.State.ErrorRet)
-		espLogger("sysOptionSet was unsuccessful: "+xmlRespon.State.ErrorRet, "error")
-		return false
-	}
-	espLogger("Max Results system setting set to: "+strconv.Itoa(newMaxResults), "debug")
-	return true
 }
 
 //processEntityClean - iterates through and processes record blocks of size defined in flag configBlockSize
@@ -345,7 +291,7 @@ func processEntityClean(entity string, chunkSize int) {
 
 	} else {
 		exitLoop := false
-		for exitLoop == false {
+		for !exitLoop {
 			AllRecordIDs := getRecordIDs(entity)
 			if len(AllRecordIDs) == 0 {
 				exitLoop = true
@@ -354,8 +300,6 @@ func processEntityClean(entity string, chunkSize int) {
 			deleteRecords(entity, AllRecordIDs)
 		}
 	}
-
-	return
 }
 
 //getLowerInt
@@ -389,7 +333,7 @@ func loadConfig() cleanerConfStruct {
 
 // espLogger -- Log to ESP
 func espLogger(message string, severity string) {
-	if configDryRun == true {
+	if configDryRun {
 		message = "[DRYRUN] " + message
 	}
 	espXmlmc.SetParam("fileName", "HornbillCleanerUtility")
@@ -418,12 +362,13 @@ func logger(s string) {
 
 	//-- Open Log File
 	f, err := os.OpenFile(logFileName, os.O_APPEND|os.O_CREATE|os.O_RDWR, 0777)
-	// don't forget to close it
-	defer f.Close()
 	if err != nil {
 		color.Red("Error Creating Log File %q: %s \n", logFileName, err)
 		os.Exit(100)
 	}
+	// don't forget to close it
+	defer f.Close()
+
 	// assign the file to the standard logger
 	log.SetOutput(f)
 	//Write the log entry